Login system rework
Overhauled existing login.html to include updated design elements, a "Remember Me" function, as well as flash messages. The sidebar in base.html was modified to include a dynamic display of a user's cash balance, as well as a logout function. The temporary logout link in profile.html was removed.